+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Persist wallet hash id
 - Click side bar to toggle collapsed
 - Add copyable label with transaction hash to the end of send flow